improve solution for asking for confirmation based on throw_error
