+++ source/main.c
@@ -316,23 +316,12 @@ int main(void)
         }
         else if(kDown & KEY_SELECT)
         {
-            while(aptMainLoop())
+            bool confirmed = draw_confirm("Are you sure you would like to delete this?", current_list, current_mode);
+            if(confirmed)
             {
-                hidScanInput();
-                u32 kDown = hidKeysDown();
-                draw_interface(current_list, current_mode);
-                pp2d_draw_on(GFX_TOP, GFX_LEFT);
-                draw_confirm("Are you sure you would like to delete this?");
-                pp2d_end_draw();
-                if(kDown & KEY_A)
-                {
-                    draw_install(INSTALL_ENTRY_DELETE);
-                    delete_entry(*current_entry);
-                    load_lists(lists);
-                    break;
-                }
-                else if(kDown & KEY_B)
-                    break;
+                draw_install(INSTALL_ENTRY_DELETE);
+                delete_entry(*current_entry);
+                load_lists(lists);
             }
         }
